+++ src/boundary.cpp
@@ -15,17 +15,17 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
     // X
     // LEFT FACE
     face = 2;
-    for (size_t k=0; k < NZ+1; k++)
+    for (int k=0; k < NZ+1; k++)
     {
         Yx[k] = boundary_value_u[face]->value(0,0,k,t);
     }
 
-    for (size_t j=1; j < NY; j++)
+    for (int j=1; j < NY; j++)
     {
         face = 4;
         Yx[j*(NZ+1)] = boundary_value_u[face]->value(0,j,0,t);
         face = 0;
-        for(size_t k = 1; k < NZ; k++)
+        for(int k = 1; k < NZ; k++)
         {
             Yx[j*(NZ+1) + k] = approximate_boundary_u(0,j,k,t,face,1);
         }
@@ -34,21 +34,21 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
     }
 
     face = 3;
-    for (size_t k=0; k < NZ+1; k++)
+    for (int k=0; k < NZ+1; k++)
     {
         Yx[NY * (NZ+1) + k] = boundary_value_u[face]->value(0,NY,k,t);
     }
 
     // MIDDLE POINTS
-    for (size_t i=1; i < NX-1; i++)
+    for (int i=1; i < NX-1; i++)
     {
         face = 2;
-        for (size_t k=0; k < NZ+1; k++)
+        for (int k=0; k < NZ+1; k++)
         {
             Yx[i * (NY+1) * (NZ+1) + k] = boundary_value_u[face]->value(i,0,k,t);
         }
         
-        for (size_t j=1; j < NY; j++)
+        for (int j=1; j < NY; j++)
         {
             face = 4;
             Yx[(NY+1) * (NZ+1) * i + j*(NZ+1)] = boundary_value_u[face]->value(i,j,0,t);
@@ -58,7 +58,7 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
         }
 
         face = 3;
-        for (size_t k=0; k < NZ+1; k++)
+        for (int k=0; k < NZ+1; k++)
         {
             Yx[NY * (NZ + 1) + (NY+1) * (NZ+1) * i + k] = boundary_value_u[face]->value(i,NY,k,t);
         }
@@ -66,18 +66,18 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
 
     // RIGHT FACE
     face = 2;
-    for (size_t k=0; k < NZ+1; k++)
+    for (int k=0; k < NZ+1; k++)
     {
         Yx[(NX-1) * (NY+1) * (NZ+1) + k] = boundary_value_u[face]->value(NX-1,0,k,t);
     }
 
-    for (size_t j=1; j < NY; j++)
+    for (int j=1; j < NY; j++)
     {
         face = 4;
         Yx[(NX-1) * (NY+1) * (NZ+1) + j*(NZ+1)] = boundary_value_u[face]->value(NX-1,j,0,t);
 
         face = 1;
-        for(size_t k = 1; k < NZ; k++)
+        for(int k = 1; k < NZ; k++)
         {
             Yx[(NX-1)*(NY+1)*(NZ+1) + j*(NZ+1) + k] = approximate_boundary_u(NX,j,k,t,face,-1);
         }
@@ -87,7 +87,7 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
     }
 
     face = 3;
-    for (size_t k=0; k < NZ+1; k++)
+    for (int k=0; k < NZ+1; k++)
     {
         Yx[(NX-1) * (NY+1) * (NZ+1) + NY*(NZ+1) + k] = boundary_value_u[face]->value(NX-1,NY,k,t);
     }
@@ -96,28 +96,28 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
     // Y
     // LEFT FACE
     face = 0;
-    for (size_t j = 0; j < NY; j++)
+    for (int j = 0; j < NY; j++)
     {
-        for(size_t k = 0; k < NZ+1; k++)
+        for(int k = 0; k < NZ+1; k++)
         {
             Yy[j*(NZ+1) + k] = boundary_value_v[face]->value(0,j,k,t);
         }
     }
 
     // MIDDLE POINTS
-    for (size_t i=1; i < NX; i++)
+    for (int i=1; i < NX; i++)
     {
         face = 4;
         Yy[NY * (NZ+1) * i] = boundary_value_v[face]->value(i,0,0,t); 
         face = 2;
-        for (size_t k=1; k < NZ; k++)
+        for (int k=1; k < NZ; k++)
         {
             Yy[i * NY * (NZ+1) + k] = approximate_boundary_v(i,0,k,t,face,1);
         }
         face = 5;
         Yy[NY * (NZ+1) * i + NZ] = boundary_value_v[face]->value(i,0,NZ,t); 
 
-        for (size_t j=1; j < NY-1; j++)
+        for (int j=1; j < NY-1; j++)
         {
             face = 4;
             Yy[NY * (NZ+1) * i + j*(NZ+1)] = boundary_value_v[face]->value(i,j,0,t); 
@@ -129,7 +129,7 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
         face = 4;
         Yy[NY * (NZ+1) * i + (NY-1)*(NZ+1)] = boundary_value_v[face]->value(i,NY-1,0,t); 
         face = 3;
-        for(size_t k = 1; k < NZ; k++)
+        for(int k = 1; k < NZ; k++)
         {
             Yy[(NY-1) * (NZ+1) + NY * (NZ+1) * i + k] = approximate_boundary_v(i,NY,k,t,face,-1);
         }
@@ -139,9 +139,9 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
 
     // RIGHT FACE
     face = 1;
-    for (size_t j = 0; j < NY; j++)
+    for (int j = 0; j < NY; j++)
     {
-        for(size_t k = 0; k < NZ+1; k++)
+        for(int k = 0; k < NZ+1; k++)
         {
             Yy[NX*NY*(NZ+1) + j*(NZ+1) + k] = boundary_value_v[face]->value(NX,j,k,t);
         }
@@ -151,24 +151,24 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
     // Z
     // LEFT FACE 
     face = 0;
-    for (size_t j = 0; j < NY+1; j++)
+    for (int j = 0; j < NY+1; j++)
     {
-        for(size_t k = 0; k < NZ; k++)
+        for(int k = 0; k < NZ; k++)
         {
             Yz[j*NZ + k] = boundary_value_w[face]->value(0,j,k,t); 
         }
     }
 
     // MIDDLE POINTS
-    for (size_t i=1; i < NX; i++)
+    for (int i=1; i < NX; i++)
     {
         face = 2;
-        for (size_t k=0; k < NZ; k++)
+        for (int k=0; k < NZ; k++)
         {
             Yz[i * (NY+1) * NZ + k] = boundary_value_w[face]->value(i,0,k,t);
         }
 
-        for(size_t j = 1; j < NY; j++)
+        for(int j = 1; j < NY; j++)
         {
             face = 4;
             Yz[(NY+1) * NZ * i + j*NZ] = approximate_boundary_w(i,j,0,t,face,1);/* boundary_value_w[face]->value(i,j,0,t); */
@@ -178,7 +178,7 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
         } 
 
         face = 3;
-        for(size_t k = 0; k < NZ+1; k++)
+        for(int k = 0; k < NZ+1; k++)
         {
             Yz[NY * NZ + i * (NY+1) * NZ + k] = boundary_value_w[face]->value(i,NY,k,t);
         }
@@ -186,9 +186,9 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
 
     // RIGHT FACE
     face = 1;
-    for (size_t j = 0; j < NY+1; j++)
+    for (int j = 0; j < NY+1; j++)
     {
-        for(size_t k = 0; k < NZ; k++)
+        for(int k = 0; k < NZ; k++)
         {
             Yz[NX*(NY+1)*NZ + j*NZ + k] = boundary_value_w[face]->value(NX,j,k,t); 
         }
@@ -205,7 +205,7 @@ void Boundary::update_boundary(std::array<Real, NX *(NY + 1) * (NZ + 1)> &Yx, st
  * 
  * @return the approximate value.
 */
-Real Boundary::approximate_boundary_u(size_t x, size_t y, size_t z, Real t, size_t face, int side)
+Real Boundary::approximate_boundary_u(int x, int y, int z, Real t, int face, int side)
 {
 
     Real dv = (boundary_value_v[face]->value(x, y, z, t) - boundary_value_v[face]->value(x, y - 1.0, z, t)) / DY;
@@ -224,7 +224,7 @@ Real Boundary::approximate_boundary_u(size_t x, size_t y, size_t z, Real t, size
  * 
  * @return the approximate value.
 */
-Real Boundary::approximate_boundary_v(size_t x, size_t y, size_t z, Real t, size_t face, int side)
+Real Boundary::approximate_boundary_v(int x, int y, int z, Real t, int face, int side)
 {
     Real du = ((boundary_value_u[face]->value(x, y, z, t)) -
                  ((boundary_value_u[face]->value(x - 1.0, y, z, t)))) /
@@ -246,7 +246,7 @@ Real Boundary::approximate_boundary_v(size_t x, size_t y, size_t z, Real t, size
  * 
  * @return the approximate value.
 */
-Real Boundary::approximate_boundary_w(size_t x, size_t y, size_t z, Real t, size_t face, int side)
+Real Boundary::approximate_boundary_w(int x, int y, int z, Real t, int face, int side)
 {
     Real du = ((boundary_value_u[face]->value(x, y, z, t)) -
                  ((boundary_value_u[face]->value(x - 1.0, y, z, t)))) /
